# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "Time": {
        "properties": {
          "isPaid": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          },
          "processes": {
            "items": {},
            "type": "array"
          },
          "keyboardKeys": {
            "type": "number"
          },
          "minutesActive": {
            "type": "number"
          },
          "mouseKeys": {
            "type": "number"
          },
          "mouseDistance": {
            "type": "number"
          },
          "fromAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "toAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "keyboardKeys",
          "minutesActive",
          "mouseKeys",
          "mouseDistance",
          "fromAt",
          "toAt"
        ],
        "example": {
          "id": "ebdf0ff6-0e3d-4381-be27-60d4a083b6cc"
        }
      },
      "Project": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "workerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "viewerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "workers": {
            "items": {},
            "type": "array"
          },
          "viewers": {
            "items": {},
            "type": "array"
          },
          "trackScreenshots": {
            "type": "boolean"
          },
          "trackProcesses": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "minLength": 1
          },
          "rateHour": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "minLength": 1
          },
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "text",
          "state"
        ],
        "example": {
          "id": "c382a47e-ed48-4700-b2df-baa89aba3f22"
        }
      },
...
      "TimeSearchSortDto": {
        "properties": {
          "createdAt": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "updatedAt": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "fromAt": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "toAt": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "keyboardKeys": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "minutesActive": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "mouseKeys": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "mouseDistance": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          }
        },
        "type": "object"
      },
...
      "TimeSearchDto": {
        "properties": {
          "sort": {
            "$ref": "#/components/schemas/TimeSearchSortDto",
            "type": "object"
          },
          "filter": {
            "$ref": "#/components/schemas/TimeSearchFilterDto",
            "type": "object"
          },
          "page": {
            "minLength": 1,
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "type": "object",
        "required": [
          "sort",
          "filter",
          "page"
        ]
      },
      "TimeCreateDto": {
        "properties": {
          "fromIndex": {
            "type": "number"
          },
          "toIndex": {
            "type": "number"
          },
          "fromAt": {
            "pattern": "\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d.\\d+Z?",
            "type": "string"
          },
          "toAt": {
            "pattern": "\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d.\\d+Z?",
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "minutesActive": {
            "type": "number"
          },
          "keyboardKeys": {
            "type": "number"
          },
          "mouseKeys": {
            "type": "number"
          },
          "mouseDistance": {
            "type": "number"
          },
          "screenshot": {
            "type": "string"
          },
          "processes": {
            "items": {},
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "fromIndex",
          "toIndex",
          "fromAt",
          "toAt",
          "projectId",
          "note",
          "minutesActive",
          "keyboardKeys",
          "mouseKeys",
          "mouseDistance"
        ]
      },
...
      "TimeIdsDto": {
        "properties": {
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array",
            "minItems": 1
          }
        },
        "type": "object",
        "required": [
          "ids"
        ]
      },
      "TimeInsertionResultDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/TimeInsertionErrorDto"
          },
          "fromIndex": {
            "type": "number"
          },
          "toIndex": {
            "type": "number"
          },
          "fromAt": {
            "pattern": "\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d.\\d+Z?",
            "type": "string"
          },
          "toAt": {
            "pattern": "\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d.\\d+Z?",
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "minutesActive": {
            "type": "number"
          },
          "keyboardKeys": {
            "type": "number"
          },
          "mouseKeys": {
            "type": "number"
          },
          "mouseDistance": {
            "type": "number"
          },
          "screenshot": {
            "type": "string"
          },
          "processes": {
            "items": {},
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "fromIndex",
          "toIndex",
          "fromAt",
          "toAt",
          "projectId",
          "note",
          "minutesActive",
          "keyboardKeys",
          "mouseKeys",
          "mouseDistance"
        ]
      },
...
      "Time_search": {
        "type": "object",
        "description": "Subset of Time serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "project": {
            "$ref": "#/components/schemas/Project_search"
          },
          "user": {
            "$ref": "#/components/schemas/User_search"
          },
          "isPaid": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          },
          "processes": {
            "items": {},
            "type": "array"
          },
          "keyboardKeys": {
            "type": "number"
          },
          "minutesActive": {
            "type": "number"
          },
          "mouseKeys": {
            "type": "number"
          },
          "mouseDistance": {
            "type": "number"
          },
          "fromAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "toAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        }
      },
      "Time_edit": {
        "type": "object",
        "description": "Subset of Time serialized with class-transformer group \"edit\".",
        "properties": {
          "isPaid": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (same value the API returns on login)"
      }
    }
  },
  "info": {
    "title": "API schema",
    "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/time/search": {
      "post": {
        "operationId": "TimeController.search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSearchDto",
                "type": "object",
                "required": [
                  "page",
                  "filter",
                  "sort"
                ],
                "properties": {
                  "page": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Zero-based page index"
                  },
                  "sort": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Map of field name to sort direction (ASC or DESC)"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Optional page size override"
                  },
                  "filter": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Endpoint-specific filters (see summary / examples)"
                  }
                }
              },
              "example": {
                "filter": {
                  "projectId": "dd8a088d-00c1-499f-a75b-7ca45821a7e3",
                  "fromAt": 1700000000000,
                  "toAt": 1700086400000
                },
                "sort": {
                  "createdAt": "ASC"
                },
                "page": 0
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Time_search"
                        }
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  }
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Search time entries",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time/paid": {
      "post": {
        "operationId": "TimeController.markPaid",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeIdsDto"
              },
              "example": {
                "ids": [
                  "dd8a088d-00c1-499f-a75b-7ca45821a7e3"
                ]
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark time entries as paid",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time/unpaid": {
      "post": {
        "operationId": "TimeController.markUnpaid",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeIdsDto"
              },
              "example": {
                "ids": [
                  "dd8a088d-00c1-499f-a75b-7ca45821a7e3"
                ]
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark time entries as unpaid",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time/totals/{id}/project": {
      "get": {
        "operationId": "TimeController.getTotals",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Aggregated time totals for project owner, workers, and viewers",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time/report/{id}": {
      "get": {
        "operationId": "TimeController.getReport",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Cached time report for a project",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time": {
      "post": {
        "operationId": "TimeController.createOrUpdateMany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "object",
                  "$ref": "#/components/schemas/TimeCreateDto"
                },
                "type": "array"
              },
              "example": [
                {
                  "fromIndex": 1000,
                  "toIndex": 2000,
                  "note": "Focus block",
                  "keyboardKeys": 3,
                  "minutesActive": 4,
                  "mouseKeys": 3,
                  "mouseDistance": 5,
                  "fromAt": "2024-01-21T09:00:00.000Z",
                  "toAt": "2024-01-21T09:10:00.000Z",
                  "projectId": "dd8a088d-00c1-499f-a75b-7ca45821a7e3"
                }
              ]
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeInsertionResultDto"
                  }
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Create or update many time rows (batch)",
        "tags": [
          "Time"
        ]
      }
    },
    "/api/time/{id}": {
      "put": {
        "operationId": "TimeController.edit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Time_edit"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Time does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Time does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Update time entry (tracking worker only)",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "TimeController.read",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Time"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Time_search"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Time does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Time does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get one time entry",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "TimeController.delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Time does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Time does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Delete time entry (tracking worker only)",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time/{id}/screenshot": {
      "delete": {
        "operationId": "TimeController.removeScreenshot",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Time does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Time does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Remove screenshot from time entry (tracking worker only)",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/time/{id}/processes": {
      "delete": {
        "operationId": "TimeController.removeProcesses",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Time does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Time does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Remove processes from time entry (tracking worker only)",
        "tags": [
          "Time"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}